promise gcc -W that we won't abuse strtok_r
gcc -W is worried that we might call strtok_r(s, delim, &var) with
uninitialized "var", because it sees that "S" may be NULL.
If that were ever to happen, then strtok_r would write through
the arbitrary pointer.  This is just stopgap in 2 of the 3 cases.
In those, we should handle strdup failure.
* proxy.c (proxy_repl_prod): Assert that strtok_r input pointer is
non-NULL.
* (proxy_repl_cons, repl_worker_del): Likewise.