correct a bug: add ":" after "f" in getopt option spec
Change --fsmode to ---fs-mode (note the three leading hyphens).
That is consistent with the fact that this option is now
internal-only and not documented.
Remove LOC_ID from Usage line, now that that is no longer
supported.
This also corrects a problem (spotted by Jeff Darcy) with my
previous change: it had made it so -f FILE no longer worked.