cure a hang on S3 error
If tabled throws an error when iwhd tries to store an object, iwhd
hangs forever. There was even a "TBD" marker about it.

An easy way to reproduce this is to store the object into a bucket
that does not exist, which is wrong but should not hang iwhd, like so:
 curl -T moo.data http://localhost:9090/nobucket/object2

Fix is to introduce a concept of consumer that failed. The number of
failed consumers is accounted as a "floor" in cons_error, so that the
producer and remaining consumers continue to function normally.

This, however makes a transfer to succeed even all replicas failed
to store the data. So, once transfer is complete, we look not just
at the error code, but also at the error count.

That's the core patch. However, if we just use pipe_cons_signal with
an error flag, the Microhttpd first returns "100 Continue" response,
and then "500 Internal Server Error". The discussion for the
100-continue suggests that it's a bad idea. We should return
any kind of error instead of "100 Continue", or otherwise
accept the data. To change that, I added one more handshake
with pipe_cons_siginit, that allows us to do this.

This was tested with basic bucket creation and object put/get
operations with s3 and fs backends.