Use gnulib
Add gnulib as a submodule.
Move ax/*.m4 to m4/*.
* bootstrap: New file.  Copied from gnulib.
Fill in blanks in bootstrap.conf template.
New file, cfg.mk, is very similar to what's used in other projects.
* configure.ac: use gl_INIT and gl_EARLY, per documentation.
Generate a Makefile for each of the new lib/ and gnulib-tests/
directories.
* Makefile.am (SUBDIRS): Update so we build in lib before "."
and so that "make check" runs the tests in gnulib-tests.
* rest.c: Do not hard-code 1.0 as version number.
* .gitignore: Ignore gnulib-tests.