[patch repod] add autostart of Mongo
We would like to start the required database instance automatically,
so that it does not need to be configured. Naturally doing is requires
certain assumptions, in particular where the database is located
(we assume that it is in a subdirectory _db/ of the current directory).

This version of the patch attempts to stop the database that we launched.