switch to atexit(auto_stop)
The main impetus for this change is that query.peg has an error() that
can exit. However, we cannot add an explicit auto_stop in front of it,
because the parser is also linked into tests, which do not have autostart
code and thus throw a link error.

Secondarily, since atexit() is only called when the start of Mongo is
assured, we do not really need checking for PID being valid or any other
scheme to disable auto_stop in the child, etc. We're not changing this
for now, but maybe later.