Fix re-replication on attribute change.
Problem #1 was that the back-end code sometimes used ms->url instead of
ms->bucket and ms->key, causing the other end to see the URL for the
attribute that was changed instead of the URL for the file being
re-replicated.  Problem #2 was a memory leak when we re-use ms->pipe
in the replica module.

backend.c: fix functional error #1
rest.c fix leak #2 by freeing ms->pipe.data_ptr in proxy_put_attr