tell GC about the thread spawned by MHD_start_daemon
rest.c's main program calls MHD_start_daemon to register
access_handler as a function that it will call from the thread
it creates. Normally, the garbage collector learns of pthread_create
calls because they're cpp-wrapped.  However, when it's called from
3rd-party libraries as in this case, we can't very well recompile,
so have to use a different approach:
* rest.c (access_handler): Call GC_register_my_thread to inform the
garbage collector of this new thread.