microhttpd may also spawn threads to call prov_list_generator; tell GC
Just as done for access_handler (registered via MHD_start_daemon),
we have to tell the garbage collector about the thread that runs
MHD_connection_handle_idle and calls prov_list_generator.
The symptom is an abort when the MHD_connection_handle_idle runs.
To diagnose, invoke gdb on the resulting core file, then type
"thread apply all bt" and note the functions at the base of the
stack on the losing thread.  That's the one that hasn't yet been
registered for GC.
* rest.c (gc_register_thread): New macro, factored out of...
(access_handler): ...here.
(prov_list_generator): Use it here, too.