build: remove unnecessary uses of xml-related cflags and ldflags
Thus, no result of the m4/libxml2.m4 test is used.  However, we'll
leave the configure-time test as well as the spec-file "BuildRequires",
since some soon-to-be-added RHEV-M-related code will use it.
* Makefile.am (iwhd_CPPFLAGS): Remove $(LIBXML_CFLAGS).
It is unnecessary.  While iwhd uses libxml2, it does so
only through hail.
(iwhd_LDADD): Remove $(LIBXML_LIBS).
Thanks to Pete Zaitcev for pointing out that $(LIBXML_CFLAGS)
was unnecessary.

2011-02-17  Jim Meyering  <meyering@redhat.com>