[patch iwhd 2/3] Fix error handling and waitpid
During the debugging of RHEV-M registration it became evident that if you
fork a child, you have to process the child's output first, and then call
waitpid(). Doing otherwise risks deadlocking if child overflows pipe.
Also, any error messages are invisible, because in case of nonzero exit
code we bail out before they are read from the pipe.

This patch adjusts s3_register to be uniform with fs_rhevm_register,
which is tested to do it right. Jeff Darcy is apprised and agrees.

Conversely, this fixes up fs_rhevm_register to use the cleanup scheme
of organ[] that s3_register has, because it is correct.

One final thing: fgets() differs from strncpy(): it always stores the
terminating nul. So, don't forcenul at [sizeof(buf)-1]. However, we
should kill the newline, or else debugging output looks ugly.

Signed-off-by: Pete Zaitcev <zaitcev@redhat.com>