Return 404 from non-existing bucket
Maros found that a GET on non-existing bucket returns an empty list
of objects, thus making it impossible to distinguish between a non-exsting
bucket and an empty bucket. This occurs because there is no object that
describes "bucket" as such in database.

However in reality we always have "special" objects, such as the
default replication policy. Therefore, we can distinguish buckets
that do not exist, and have no objects of any kind, and those that
are empty, but actually have "_default" object.

One small difficulty is that the HTTP code has to be known before
the data transfer begins. For this to work, this patch starts the
query before MicroHTTPd response is set up. If we find any keys,
including hidden ones, the bucket is considered existing. Then,
the just-found key is saved for the proxy_query_func to return.
If the query returns nothing, we return 404 to the user.