Make directories that we need in export domain.
We found by trial that RHEV-M creates any directories it needs when they
are needed, but not before. So, a fresh export domain that never hosted
any templates would not have UUID/master, and we need it. The fix is to
create that. Note that we do not create UUID/images, because that is
made when the domain is attached, and one cannot register to an unattached
domain anyhow (because one cannot predict UUID of it).