build: check for mongodb headers
On my Fedora 14 system, without mongodb-devel I was seeing:

      CXX    meta.o
    meta.cpp:34:35: fatal error: mongo/client/dbclient.h: No such file or directory
    compilation terminated.

The header is C++ so we need AC_LANGUAGE_PUSH and, similar to meta.cpp,
we need to undefine VERSION before including the header.

Note, the AC_CHECK_LIB approach doesn't seem sufficient because the
mongodb package in Fedora doesn't ship versioned shared libraries with an
unversioned symlink in the -devel package.

With the new check, we fail with:

  checking for mongo/client/dbclient.h... no
  configure: error: Missing Mongo DB client development library: mongodb-devel