Add registration in vSphere (with Jim's fix-ups)
This patch adds yet another cloud type: VMware vSphere. The new backend
type is "fs-vmw". The RPC calls are hand-rolled with libxml2 and libcurl,
so new new dependencies are necessary. Documentation is updated.

The implementation is not perfect. Two major known issues are:
 - this is not endian clean
 - reported completion percentages are fake
 - error reporting is crying for stderr captures in backend.c

However, it is desired to get something working into hands of Aeolus
people as soon as possible. The issues are to be addressed.