iwhd: avoid segfault during garbage collection
Any function that is called in its own thread
via MHD_create_response_from_callback must begin with a call to
gc_register_thread, and end with a call to GC_unregister_my_thread.
For each of the following functions, since all but one has multiple
return points, create a trivial wrapper function with the same name
as the original function and rename the original to ORIGINAL_0.
The trivial wrapper ensures that the register/unregister calls
are made and returns the result.
* rest.c (proxy_get_cons): As above.
(proxy_query_func): Likewise.
(root_blob_generator): Likewise.
(parts_callback): Likewise.
* NEWS (Bug fixes): Mention it.