Add RHEV-M template import
Most clouds structure the procedure to start a VM into 2 general steps:
1. Upload and "register" the bundle, so that it is known to the cloud.
This typically returns some kind of ID, such as Amazon's ami-xxxxxxxx.
2. Instantiate the "registered" image so that a running VM is created.

Unlike other clouds, RHEV-M requires a 3-step process to start a VM:
1. Upload the bundle into so-called "export domain". This makes a
"template" known to the cloud, but it cannot be launched yet.
2. Import the template. At its core it means copying the template
from "export domain" to "master domain".
3. Instantiate the imported template to produce VMs.

One way or the other, someone has to do this: either iwhd itself,
or iwhd's user (Aeolus Conductor). This patch implements the middle
step in iwhd.

See the update to registration.md for the new parameter "cluster".